Add test cases for new, suggested behavior
The Gherkin reference specifically stipulates that keywords should not
be taken into account when looking for a step definition [1].

I've personally witnessed in several projects that this can lead to
deterioration in test language unless every change is strictly reviewed.

I propose a stricter lookup, which does take keyword into account when
looking for a step definition, but not at the expense of ambiguity in
the test language.

I'm merely suggesting something that guards us from accidentally using
the wrong keyword in the tests.

[1] https://cucumber.io/docs/gherkin/reference/#steps